refactor: move category ui state to shared
