+++ presentation/src/main/java/com/madrid/presentation/screens/homeScreen/layout/CategoriesLayout.kt
@@ -32,8 +32,8 @@ import com.madrid.designSystem.modifier.ShimmerCard
 import com.madrid.designSystem.modifier.removeWidthPaddingFromParent
 import com.madrid.presentation.R
 import com.madrid.presentation.component.movioCards.MovioVerticalCard
-import com.madrid.presentation.viewModel.homeViewModel.CategoryUiState
 import com.madrid.presentation.viewModel.homeViewModel.SortingType
+import com.madrid.presentation.viewModel.shared.CategoryUiState
 import com.madrid.presentation.viewModel.shared.MediaType
 import com.madrid.presentation.viewModel.shared.MediaUiState
