+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/movies/SeeAllMoviesInteractionListener.kt
@@ -1,5 +1,7 @@
 package com.madrid.presentation.viewModel.seeAll.movies
 
+import com.madrid.presentation.viewModel.shared.CategoryUiState
+
 interface SeeAllMoviesInteractionListener {
     fun onMovieClick(movieId:Int)
     fun onGenreSelect(genre: CategoryUiState? = null)
