+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/movies/SeeAllMoviesUiState.kt
@@ -2,6 +2,7 @@ package com.madrid.presentation.viewModel.seeAll.movies
 
 import androidx.paging.PagingData
 import com.madrid.domain.entity.Genre
+import com.madrid.presentation.viewModel.shared.CategoryUiState
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
@@ -21,15 +22,3 @@ data class MoviesUiState(
     val name: String = "",
     val genre: List<CategoryUiState> = emptyList()
 )
-
-data class CategoryUiState(
-    val id: Int = 0,
-    val name: String = ""
-)
-
-fun Genre.toCategoryUiState(): CategoryUiState {
-    return CategoryUiState(
-        id = this.id,
-        name = this.name
-    )
-}
