+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/tvShows/SeeAllTVShowsInteractionListener.kt
@@ -1,5 +1,7 @@
 package com.madrid.presentation.viewModel.seeAll.tvShows
 
+import com.madrid.presentation.viewModel.shared.CategoryUiState
+
 interface SeeAllTVShowsInteractionListener {
     fun onSeriesClick(seriesId:Int)
     fun onGenreSelect(genre: CategoryUiState? = null)
