+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/tvShows/SeeAllTVShowsUiState.kt
@@ -1,6 +1,7 @@
 package com.madrid.presentation.viewModel.seeAll.tvShows
 
 import androidx.paging.PagingData
+import com.madrid.presentation.viewModel.shared.CategoryUiState
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
@@ -19,9 +20,4 @@ data class SeriesUiState(
     val rate: String = "",
     val name: String = "",
     val genre: List<CategoryUiState> = emptyList()
-)
-
-data class CategoryUiState(
-    val id: Int = 0,
-    val name: String = ""
 )
