+++ presentation/src/main/java/com/madrid/presentation/viewModel/shared/CategoryUiState.kt
@@ -1,0 +1,15 @@
+package com.madrid.presentation.viewModel.shared
+
+import com.madrid.domain.entity.Genre
+
+
+data class CategoryUiState(
+    val id: Int = -1,
+    val name: String = "All",
+)
+
+fun Genre.toCategoryUiState() = CategoryUiState(
+    id = this.id,
+    name = this.name
+)
+
